Check also QSL-Images (like u did for eQSL) b4 delete

+++ application/models/Qsl_model.php
@@ -53,15 +53,19 @@ class Qsl_model extends CI_Model {
 	function del_image_for_qso($qso_id, $user_id = null) {
 		// QSO belongs to station_profile. But since we have folders for Users (and therefore an extra indirect relation) we need to lookup user for station first...
 		$qsl_img=$this->db->query('SELECT e.filename, e.id, qso.station_id, s.user_id FROM qsl_images e INNER JOIN '.$this->config->item('table_name').' qso ON (e.qsoid = qso.COL_PRIMARY_KEY) inner join station_profile s on (s.station_id=qso.station_id) where qso.COL_PRIMARY_KEY=?',$qso_id);
+
+		if (!valid_uid($user_id)) {
+			$user_id = $this->session->userdata('user_id');
+		}
 		foreach ($qsl_img->result() as $row) {
-			if (($user_id ?? '') == '') {					// Calling as User? Check if User-id matches User-id from QSO
-				$user_id = $this->session->userdata('user_id');
-				if ($row->user_id != $user_id) {
-					return "No Image";				// Image doesn't belong to user, so return
-				}
+			// Calling as User? Check if User-id matches User-id from QSO
+			if ($row->user_id != $user_id) {
+				return "No Image"; // Image doesn't belong to user, so return
 			}
-			$image = $this->paths->getUserdataPath('qsl_card', 'p',$row->user_id).'/'.$row->filename;
-			unlink($image);
+			$image = $this->paths->getUserdataPath('qsl_card', 'p', $row->user_id).'/'.$row->filename;
+			if (file_exists($image)) {
+				unlink($image);
+			}
 			$this->db->delete('qsl_images', array('id' => $row->id));
 		}
 	}
@@ -83,7 +87,9 @@ class Qsl_model extends CI_Model {
 		$path = $this->paths->getUserdataPath('qsl_card', 'p');
 		$file = $this->getFilename($clean_id)->row();
 		$filename = basename($file->filename);
-		unlink($path.'/'.$filename);
+		if (file_exists($path.'/'.$filename)) {
+			unlink($path.'/'.$filename);
+		}
 		// Delete Mode
 		$this->db->delete('qsl_images', array('id' => $clean_id));
 	}
